python-services/data-processing/requirements.txt
fastapi
uvicorn[standard]
fastf1
python-dotenv